Improve readiness of cmd structure
